backend/requirements.txt
# ============================================
# CORE FRAMEWORK
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2

# ============================================
# VALIDATION
# ============================================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0

# ============================================
# DEVELOPMENT (Optional)
# ============================================
# pytest==7.4.4
# httpx==0.26.0